+++ src/toolkit/position.cc
@@ -1,8 +1,9 @@
+#include "position.h"
+
 #include <math.h>
 #include <stdio.h>
+#include <iomanip>
 #include <sstream>
-
-#include "position.h"
 
 namespace cyclus {
 namespace toolkit {
@@ -43,8 +44,18 @@ void Position::set_position(double lat, double lon) {
   longitude_ = SetPrecision(lon * CYCLUS_DECIMAL_SECOND_MULTIPLIER, 1);
 }
 
+void Position::RecordPosition(Agent* agent) {
+  agent->context()
+      ->NewDatum("AgentPosition")
+      ->AddVal("Spec", agent->spec())
+      ->AddVal("Prototype", agent->prototype())
+      ->AddVal("AgentId", agent->id())
+      ->AddVal("Latitude", latitude_ / CYCLUS_DECIMAL_SECOND_MULTIPLIER)
+      ->AddVal("Longitude", longitude_ / CYCLUS_DECIMAL_SECOND_MULTIPLIER)
+      ->Record();
+}
 void Position::LatCheck(double lat) {
-  if (lat > 90 || lat < -90){
+  if (lat > 90 || lat < -90) {
     std::stringstream msg;
     msg << "The provided latitude (" << lat
         << ") is outside the acceptable range. "
@@ -54,7 +65,7 @@ void Position::LatCheck(double lat) {
 }
 
 void Position::LonCheck(double lon) {
-  if (lon > 180 || lon < -180){
+  if (lon > 180 || lon < -180) {
     std::stringstream msg;
     msg << "The provided longitude (" << lon
         << ") is outside the acceptable range."